requirements.txt



# google integrations
google-api-python-client==2.49.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1